EhuBio - Changed FragmentArray precission from float to double
MyMRM
- Only best (less ppm) fragment ion is used for each m/z

+++ trunk/Projects/java/EhuBio/src/es/ehubio/proteomics/psi/mzid11/FragmentArrayType.java
@@ -41,7 +41,7 @@ import javax.xml.bind.annotation.XmlType;
 public class FragmentArrayType {
 
     @XmlAttribute(name = "values", required = true)
-    protected List<Float> values;
+    protected List<Double> values;
     @XmlAttribute(name = "measure_ref", required = true)
     protected String measureRef;
 
@@ -67,9 +67,9 @@ public class FragmentArrayType {
      * 
      * 
      */
-    public List<Float> getValues() {
+    public List<Double> getValues() {
         if (values == null) {
-            values = new ArrayList<Float>();
+            values = new ArrayList<Double>();
         }
         return this.values;
     }
